refactor: move sequencer test consts to test fixtures
